fix(stock): compute past refills and float quantities (#22)

+++ backend/internal/domain/entry.go
@@ -1,7 +1,7 @@
 package domain
 
 type CreateStockEntryRequest struct {
-	Quantity int    `json:"quantity"`
-	Unit     string `json:"unit"` // "pill" or "box"
-	Date     string `json:"date"` // "2025-06-02"
+	Quantity float64 `json:"quantity"`
+	Unit     string  `json:"unit"` // "pill" or "box"
+	Date     string  `json:"date"` // "2025-06-02"
 }
